SocialMediaBot/requirements.txt
azure-functions
beautifulsoup4
tensorflow
transformers